@@ -225,10 +225,9 @@ class _SelectTemplateState extends ConsumerState<SelectSoundsScreen> {
       ),
     );
 
-    final scaffold = WillPopScope(
-      onWillPop: () async {
+    final scaffold = PopScope(
+      onPopInvoked: (_) async {
         await ref.read(widget.viewModelProvider.notifier).beforeHideScreen();
-        return true;
       },
       child: Scaffold(
         appBar: transparentAppBar(
